+++ R/summarizor.R
@@ -8,7 +8,7 @@
 #' @param x dataset
 #' @param by columns names to be used as grouping columns
 #' @param overall_label label to use as overall label
-#' @seealso [fmt_2stats()]
+#' @seealso [fmt_2stats()], [labelizor()]
 #' @examples
 #' z <- summarizor(CO2[-c(1, 4)],
 #'   by = "Treatment",
@@ -125,7 +125,7 @@ summarizor <- function(
   }
   dat <- rbindlist(dat$data)
 
-  first_levels <- c("n", "mean_sd", "median", "range")
+  first_levels <- c("n", "mean_sd", "median_iqr", "range")
   last_levels <- c("missing")
   levs <- c(first_levels, setdiff(unique(dat$stat), c(first_levels, last_levels)), last_levels)
   labs <- levs
@@ -134,6 +134,7 @@ summarizor <- function(
   dat
 }
 
+#' @importFrom stats sd IQR median
 dataset_describe <- function(dataset){
 
   w <- lapply(dataset, function(x){
@@ -149,9 +150,9 @@ dataset_describe <- function(dataset){
     } else if(is.numeric(x)){
 
       z <- data.frame(
-        stat = c("mean_sd", "median", "range", "missing"),
-        value1 = c(mean(x, na.rm = TRUE), as.double(quantile(x, probs = .5, na.rm = TRUE)), min(x, na.rm = TRUE), NA_real_),
-        value2 = c(sd(x, na.rm = TRUE), NA_real_, max(x, na.rm = TRUE), NA_real_),
+        stat = c("mean_sd", "median_iqr", "range", "missing"),
+        value1 = c(mean(x, na.rm = TRUE), as.double(median(x, na.rm = TRUE)), min(x, na.rm = TRUE), NA_real_),
+        value2 = c(sd(x, na.rm = TRUE), as.double(IQR(x, na.rm = TRUE)), max(x, na.rm = TRUE), NA_real_),
         cts = NA_real_, percent = NA_real_)
 
       z$cts[z$stat %in% "missing"] <- sum(is.na(x))
@@ -196,6 +197,7 @@ fmt_2stats <- function(stat, num1, num2, cts, pcts,
   z_cts <- character(length = length(num1))
 
   is_mean_sd <- !is.na(num1) & !is.na(num2) & stat %in% "mean_sd"
+  is_median_iqr <- !is.na(num1) & !is.na(num2) & stat %in% "median_iqr"
   is_range <- !is.na(num1) & !is.na(num2) & stat %in% "range"
   is_num_1 <- !is.na(num1) & is.na(num2)
   is_cts_2 <- !is.na(cts) & !is.na(pcts)
@@ -207,6 +209,11 @@ fmt_2stats <- function(stat, num1, num2, cts, pcts,
     sprintf(num1_mask, num1[is_mean_sd]),
     " ",
     sprintf(num2_mask, num2[is_mean_sd])
+  )
+  z_num[is_median_iqr] <- paste0(
+    sprintf(num1_mask, num1[is_median_iqr]),
+    " ",
+    sprintf(num2_mask, num2[is_median_iqr])
   )
   z_num[is_range] <- paste0(
     sprintf(num1_mask, num1[is_range]),